+++ SymCryptProvider/src/signature/p_scossl_rsa_signature.c
@@ -7,45 +7,13 @@
 #include <openssl/evp.h>
 #include <openssl/proverr.h>
 
-#include "scossl_rsa.h"
-#include "p_scossl_rsa.h"
-#include "p_scossl_base.h"
+#include "p_scossl_rsa_signature.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-typedef struct
-{
-    SCOSSL_PROV_RSA_KEY_CTX *keyCtx;
-    UINT8 padding;
-    int operation;
-
-    // Needed for fetching md
-    OSSL_LIB_CTX *libctx;
-    char* propq;
-
-    EVP_MD_CTX *mdctx;
-    EVP_MD *md;
-    const OSSL_ITEM *mdInfo; // Informational, must match md if set
-    BOOL allowMdUpdates;
-
-    // PSS params
-    BOOL pssRestricted;
-    const OSSL_ITEM *mgf1MdInfo; // Informational, must match md if set
-    int cbSalt;
-    int cbSaltMin;
-
-    // Sigalg state tracking
-    BOOL isSigalg;
-    BOOL allowUpdate;
-    BOOL allowFinal;
-    BOOL allowOneshot;
-
-    // Sigalg verify message support
-    PBYTE pbSignature;
-    SIZE_T cbSignature;
-} SCOSSL_RSA_SIGN_CTX;
+static SCOSSL_STATUS p_scossl_rsa_set_ctx_params(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, _In_ const OSSL_PARAM params[]);
 
 #define SCOSSL_RSA_SIGNATURE_GETTABLE_PARAMS                        \
     OSSL_PARAM_utf8_string(OSSL_SIGNATURE_PARAM_DIGEST, NULL, 0),   \
@@ -70,10 +38,6 @@ static const OSSL_PARAM p_scossl_rsa_pss_sig_ctx_gettable_param_types[] = {
     SCOSSL_RSA_PSS_SIGNATURE_GETTABLE_PARAMS
     OSSL_PARAM_END};
 
-static const OSSL_PARAM p_scossl_rsa_sigalg_ctx_settable_param_types[] = {
-    OSSL_PARAM_octet_string(OSSL_SIGNATURE_PARAM_SIGNATURE, NULL, 0),
-    OSSL_PARAM_END};
-
 // Padding may not be set at the time of querying settable params, so PSS params
 // are always accepted. The provider will check the padding before attempting
 // to set the PSS parameters
@@ -95,9 +59,8 @@ static OSSL_ITEM p_scossl_rsa_sign_padding_modes[] = {
     {RSA_PKCS1_PSS_PADDING, OSSL_PKEY_RSA_PAD_MODE_PSS},
     {0, NULL}};
 
-static SCOSSL_STATUS p_scossl_rsa_set_ctx_params(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, _In_ const OSSL_PARAM params[]);
-
-static SCOSSL_RSA_SIGN_CTX *p_scossl_rsa_newctx(_In_ SCOSSL_PROVCTX *provctx, _In_ const char *propq)
+_Use_decl_annotations_
+SCOSSL_RSA_SIGN_CTX *p_scossl_rsa_newctx(SCOSSL_PROVCTX *provctx, const char *propq)
 {
     SCOSSL_RSA_SIGN_CTX *ctx = OPENSSL_zalloc(sizeof(SCOSSL_RSA_SIGN_CTX));
     if (ctx != NULL)
@@ -118,7 +81,8 @@ static SCOSSL_RSA_SIGN_CTX *p_scossl_rsa_newctx(_In_ SCOSSL_PROVCTX *provctx, _I
     return ctx;
 }
 
-static void p_scossl_rsa_freectx(SCOSSL_RSA_SIGN_CTX *ctx)
+_Use_decl_annotations_
+void p_scossl_rsa_freectx(SCOSSL_RSA_SIGN_CTX *ctx)
 {
     if (ctx == NULL)
         return;
@@ -130,7 +94,8 @@ static void p_scossl_rsa_freectx(SCOSSL_RSA_SIGN_CTX *ctx)
     OPENSSL_free(ctx);
 }
 
-static SCOSSL_RSA_SIGN_CTX *p_scossl_rsa_dupctx(_In_ SCOSSL_RSA_SIGN_CTX *ctx)
+_Use_decl_annotations_
+SCOSSL_RSA_SIGN_CTX *p_scossl_rsa_dupctx(SCOSSL_RSA_SIGN_CTX *ctx)
 {
     SCOSSL_RSA_SIGN_CTX *copyCtx = OPENSSL_zalloc(sizeof(SCOSSL_RSA_SIGN_CTX));
     if (copyCtx != NULL)
@@ -165,8 +130,9 @@ static SCOSSL_RSA_SIGN_CTX *p_scossl_rsa_dupctx(_In_ SCOSSL_RSA_SIGN_CTX *ctx)
     return copyCtx;
 }
 
-static SCOSSL_STATUS p_scossl_rsa_signverify_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,
-                                                  _In_ const OSSL_PARAM params[], int operation)
+_Use_decl_annotations_
+SCOSSL_STATUS p_scossl_rsa_signverify_init(SCOSSL_RSA_SIGN_CTX *ctx, SCOSSL_PROV_RSA_KEY_CTX *keyCtx,
+                                           const OSSL_PARAM params[], int operation)
 {
     if (ctx == NULL ||
         (keyCtx == NULL && ctx->keyCtx == NULL))
@@ -252,9 +218,10 @@ static SCOSSL_STATUS p_scossl_rsa_verify_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, 
     return p_scossl_rsa_signverify_init(ctx, keyCtx, params, EVP_PKEY_OP_VERIFY);
 }
 
-static SCOSSL_STATUS p_scossl_rsa_sign_internal(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
-                                                _Out_writes_bytes_(*siglen) unsigned char *sig, _Out_ size_t *siglen, size_t sigsize,
-                                                _In_reads_bytes_(tbslen) const unsigned char *tbs, size_t tbslen)
+_Use_decl_annotations_
+SCOSSL_STATUS p_scossl_rsa_sign_internal(SCOSSL_RSA_SIGN_CTX *ctx,
+                                         unsigned char *sig, size_t *siglen, size_t sigsize,
+                                         const unsigned char *tbs, size_t tbslen)
 {
     int mdnid = ctx->mdInfo == NULL ? NID_undef : ctx->mdInfo->id;
     SCOSSL_STATUS ret = SCOSSL_FAILURE;
@@ -301,9 +268,10 @@ err:
     return ret;
 }
 
-static SCOSSL_STATUS p_scossl_rsa_verify_internal(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
-                                                  _In_reads_bytes_(siglen) const unsigned char *sig, size_t siglen,
-                                                  _In_reads_bytes_(tbslen) const unsigned char *tbs, size_t tbslen)
+_Use_decl_annotations_
+SCOSSL_STATUS p_scossl_rsa_verify_internal(SCOSSL_RSA_SIGN_CTX *ctx,
+                                           const unsigned char *sig, size_t siglen,
+                                           const unsigned char *tbs, size_t tbslen)
 {
     int mdnid = ctx->mdInfo == NULL ? NID_undef : ctx->mdInfo->id;
 
@@ -506,6 +474,44 @@ static SCOSSL_STATUS p_scossl_rsa_digest_verify_final(_In_ SCOSSL_RSA_SIGN_CTX *
     ctx->allowMdUpdates = TRUE;
 
     return ret;
+}
+
+static SCOSSL_STATUS p_scossl_rsa_sign(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
+                                       _Out_writes_bytes_(*siglen) unsigned char *sig, _Out_ size_t *siglen, size_t sigsize,
+                                       _In_reads_bytes_(tbslen) const unsigned char *tbs, size_t tbslen)
+{
+    if (ctx == NULL)
+    {
+        ERR_raise(ERR_LIB_PROV, PROV_R_NO_KEY_SET);
+        return SCOSSL_FAILURE;
+    }
+
+    if (!ctx->allowOneshot)
+    {
+        ERR_raise(ERR_LIB_PROV, PROV_R_ONESHOT_CALL_OUT_OF_ORDER);
+        return SCOSSL_FAILURE;
+    }
+
+    return p_scossl_rsa_sign_internal(ctx, sig, siglen, sigsize, tbs, tbslen);
+}
+
+static SCOSSL_STATUS p_scossl_rsa_verify(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
+                                         _In_reads_bytes_(siglen) const unsigned char *sig, size_t siglen,
+                                         _In_reads_bytes_(tbslen) const unsigned char *tbs, size_t tbslen)
+{
+    if (ctx == NULL)
+    {
+        ERR_raise(ERR_LIB_PROV, PROV_R_NO_KEY_SET);
+        return SCOSSL_FAILURE;
+    }
+
+    if (!ctx->allowOneshot)
+    {
+        ERR_raise(ERR_LIB_PROV, PROV_R_ONESHOT_CALL_OUT_OF_ORDER);
+        return SCOSSL_FAILURE;
+    }
+
+    return p_scossl_rsa_verify_internal(ctx, sig, siglen, tbs, tbslen);
 }
 
 static const OSSL_PARAM *p_scossl_rsa_settable_ctx_params(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
@@ -777,42 +783,8 @@ static SCOSSL_STATUS p_scossl_rsa_set_ctx_params(_Inout_ SCOSSL_RSA_SIGN_CTX *ct
     return SCOSSL_SUCCESS;
 }
 
-static const OSSL_PARAM *p_scossl_rsa_sigalg_settable_ctx_params(_In_ SCOSSL_RSA_SIGN_CTX *ctx, ossl_unused void *provctx)
-{
-    if (ctx != NULL && ctx->operation == EVP_PKEY_OP_VERIFYMSG)
-    {
-        return p_scossl_rsa_sigalg_ctx_settable_param_types;
-    }
-
-    return NULL;
-}
-
-static SCOSSL_STATUS p_scossl_rsa_sigalg_set_ctx_params(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, _In_ const OSSL_PARAM params[])
-{
-    const OSSL_PARAM *p;
-
-    if (ctx == NULL)
-        return SCOSSL_FAILURE;
-
-    if (ctx->operation == EVP_PKEY_OP_VERIFYMSG)
-    {
-        if ((p = OSSL_PARAM_locate_const(params, OSSL_SIGNATURE_PARAM_SIGNATURE)) != NULL)
-        {
-            OPENSSL_free(ctx->pbSignature);
-            ctx->pbSignature = NULL;
-            ctx->cbSignature = 0;
-            if (!OSSL_PARAM_get_octet_string(p, (void **)&ctx->pbSignature, 0, &ctx->cbSignature))
-            {
-                return SCOSSL_FAILURE;
-            }
-        }
-    }
-
-    return SCOSSL_SUCCESS;
-}
-
-static const OSSL_PARAM *p_scossl_rsa_gettable_ctx_params(_In_ SCOSSL_RSA_SIGN_CTX  *ctx,
-                                                          ossl_unused void *provctx)
+_Use_decl_annotations_
+const OSSL_PARAM *p_scossl_rsa_gettable_ctx_params(SCOSSL_RSA_SIGN_CTX *ctx, ossl_unused void *provctx)
 {
     return ctx->padding == RSA_PKCS1_PSS_PADDING ? p_scossl_rsa_pss_sig_ctx_gettable_param_types : p_scossl_rsa_sig_ctx_gettable_param_types;
 }
@@ -915,7 +887,8 @@ cleanup:
     return pssParamSeq;
 }
 
-static SCOSSL_STATUS p_scossl_rsa_get_ctx_params(_In_ SCOSSL_RSA_SIGN_CTX *ctx, _Inout_ OSSL_PARAM params[])
+_Use_decl_annotations_
+SCOSSL_STATUS p_scossl_rsa_get_ctx_params(SCOSSL_RSA_SIGN_CTX *ctx, OSSL_PARAM params[])
 {
     if (params == NULL)
     {
@@ -1172,221 +1145,6 @@ static SCOSSL_STATUS p_scossl_rsa_set_ctx_md_params(_In_ SCOSSL_RSA_SIGN_CTX *ct
     return EVP_MD_CTX_set_params(ctx->mdctx, params);
 }
 
-static SCOSSL_STATUS p_scossl_rsa_sigalg_signverify_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,
-                                                         _In_ const OSSL_PARAM params[], int operation,
-                                                         _In_ const char *mdname)
-{
-    EVP_MD *md = NULL;
-
-    if (!p_scossl_rsa_signverify_init(ctx, keyCtx, params, operation))
-    {
-        return SCOSSL_FAILURE;
-    }
-
-    // Sigalgs are not supported for PSS keys
-    if (ctx->padding == RSA_PKCS1_PSS_PADDING)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_OPERATION_NOT_SUPPORTED_FOR_THIS_KEYTYPE);
-        return SCOSSL_FAILURE;
-    }
-
-    const OSSL_ITEM *mdInfo = p_scossl_rsa_get_supported_md(ctx->libctx, RSA_PKCS1_PADDING, mdname, NULL, &md);
-    if (mdInfo == NULL)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_INVALID_DIGEST);
-        return SCOSSL_FAILURE;
-    }
-
-    EVP_MD_free(ctx->md);
-    ctx->md = md;
-    ctx->mdInfo = mdInfo;
-    ctx->padding = RSA_PKCS1_PADDING;
-    ctx->allowMdUpdates = FALSE;
-    ctx->isSigalg = TRUE;
-
-    return SCOSSL_SUCCESS;
-}
-
-static SCOSSL_STATUS p_scossl_rsa_sigalg_message_signverify_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx, _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,
-                                                                 _In_ const OSSL_PARAM params[], int operation,
-                                                                 _In_ const char *mdname)
-{
-    if (!p_scossl_rsa_sigalg_signverify_init(ctx, keyCtx, params, operation, mdname))
-    {
-        return SCOSSL_FAILURE;
-    }
-
-    if (ctx->mdctx == NULL)
-    {
-        if ((ctx->mdctx = EVP_MD_CTX_new()) == NULL)
-        {
-            ERR_raise(ERR_LIB_PROV, ERR_R_MALLOC_FAILURE);
-            return SCOSSL_FAILURE;
-        }
-    }
-
-    if (!EVP_DigestInit_ex2(ctx->mdctx, ctx->md, params))
-    {
-        EVP_MD_CTX_free(ctx->mdctx);
-        ctx->mdctx = NULL;
-        return SCOSSL_FAILURE;
-    }
-
-    return SCOSSL_SUCCESS;
-}
-
-static SCOSSL_STATUS p_scossl_rsa_sigalg_signverify_message_update(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
-                                                                   _In_reads_bytes_(datalen) const unsigned char *data,
-                                                                   size_t datalen)
-{
-    if (ctx == NULL || ctx->mdctx == NULL)
-    {
-        ERR_raise(ERR_LIB_PROV, ERR_R_PASSED_NULL_PARAMETER);
-        return SCOSSL_FAILURE;
-    }
-
-    if (!ctx->allowUpdate)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_UPDATE_CALL_OUT_OF_ORDER);
-        return SCOSSL_FAILURE;
-    }
-
-    ctx->allowOneshot = 0;
-
-    return EVP_DigestUpdate(ctx->mdctx, data, datalen);
-}
-
-static SCOSSL_STATUS p_scossl_rsa_sigalg_sign_message_final(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
-                                                            _Out_writes_bytes_opt_(*siglen) unsigned char *sig,
-                                                            _Out_ size_t *siglen, size_t sigsize)
-{
-    BYTE abDigest[EVP_MAX_MD_SIZE];
-    unsigned int cbDigest = 0;
-
-    if (ctx == NULL || ctx->mdctx == NULL)
-    {
-        ERR_raise(ERR_LIB_PROV, ERR_R_PASSED_NULL_PARAMETER);
-        return SCOSSL_FAILURE;
-    }
-
-    if (!ctx->allowFinal)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_FINAL_CALL_OUT_OF_ORDER);
-        return SCOSSL_FAILURE;
-    }
-
-    // If sig is NULL, this is a size fetch, and the digest does not need to be computed
-    if (sig != NULL)
-    {
-        if (!EVP_DigestFinal_ex(ctx->mdctx, abDigest, &cbDigest))
-        {
-            return SCOSSL_FAILURE;
-        }
-
-        ctx->allowUpdate = FALSE;
-        ctx->allowFinal = FALSE;
-        ctx->allowOneshot = FALSE;
-    }
-
-    return p_scossl_rsa_sign_internal(ctx, sig, siglen, sigsize, abDigest, cbDigest);
-}
-
-static int p_scossl_rsa_sigalg_verify_message_final(_In_ SCOSSL_RSA_SIGN_CTX *ctx)
-{
-    BYTE abDigest[EVP_MAX_MD_SIZE];
-    unsigned int cbDigest = 0;
-
-    if (ctx == NULL || ctx->mdctx == NULL)
-    {
-        ERR_raise(ERR_LIB_PROV, ERR_R_PASSED_NULL_PARAMETER);
-        return SCOSSL_FAILURE;
-    }
-
-    if (!ctx->allowFinal)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_FINAL_CALL_OUT_OF_ORDER);
-        return SCOSSL_FAILURE;
-    }
-
-    if (!EVP_DigestFinal_ex(ctx->mdctx, abDigest, &cbDigest))
-    {
-        return SCOSSL_FAILURE;
-    }
-
-    ctx->allowUpdate = FALSE;
-    ctx->allowFinal = FALSE;
-    ctx->allowOneshot = FALSE;
-
-    return p_scossl_rsa_verify_internal(ctx, ctx->pbSignature, ctx->cbSignature, abDigest, cbDigest);
-}
-
-static SCOSSL_STATUS p_scossl_rsa_sign(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
-                                       _Out_writes_bytes_(*siglen) unsigned char *sig, _Out_ size_t *siglen, size_t sigsize,
-                                       _In_reads_bytes_(tbslen) const unsigned char *tbs, size_t tbslen)
-{
-    if (ctx == NULL)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_NO_KEY_SET);
-        return SCOSSL_FAILURE;
-    }
-
-    if (!ctx->allowOneshot)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_ONESHOT_CALL_OUT_OF_ORDER);
-        return SCOSSL_FAILURE;
-    }
-
-    if (ctx->operation == EVP_PKEY_OP_SIGNMSG)
-    {
-        if (sig == NULL)
-        {
-            return p_scossl_rsa_sigalg_sign_message_final(ctx, sig, siglen, sigsize);
-        }
-
-        return p_scossl_rsa_sigalg_signverify_message_update(ctx, tbs, tbslen) &&
-               p_scossl_rsa_sigalg_sign_message_final(ctx, sig, siglen, sigsize);
-    }
-
-    return p_scossl_rsa_sign_internal(ctx, sig, siglen, sigsize, tbs, tbslen);
-}
-
-// Dispatch-facing verify function.
-// If verifying a message, digests tbs and verifies the result.
-// Otherwise, verifies tbs directly as a pre-computed digest.
-static SCOSSL_STATUS p_scossl_rsa_verify(_In_ SCOSSL_RSA_SIGN_CTX *ctx,
-                                         _In_reads_bytes_(siglen) const unsigned char *sig, size_t siglen,
-                                         _In_reads_bytes_(tbslen) const unsigned char *tbs, size_t tbslen)
-{
-    if (ctx == NULL)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_NO_KEY_SET);
-        return SCOSSL_FAILURE;
-    }
-
-    if (!ctx->allowOneshot)
-    {
-        ERR_raise(ERR_LIB_PROV, PROV_R_ONESHOT_CALL_OUT_OF_ORDER);
-        return SCOSSL_FAILURE;
-    }
-
-    if (ctx->operation == EVP_PKEY_OP_VERIFYMSG)
-    {
-        OPENSSL_free(ctx->pbSignature);
-        ctx->cbSignature = 0;
-        if ((ctx->pbSignature = OPENSSL_memdup(sig, siglen)) == NULL)
-        {
-            ERR_raise(ERR_LIB_PROV, ERR_R_MALLOC_FAILURE);
-            return SCOSSL_FAILURE;
-        }
-        ctx->cbSignature = siglen;
-
-        return p_scossl_rsa_sigalg_signverify_message_update(ctx, tbs, tbslen) &&
-               p_scossl_rsa_sigalg_verify_message_final(ctx);
-    }
-
-    return p_scossl_rsa_verify_internal(ctx, sig, siglen, tbs, tbslen);
-}
-
 const OSSL_DISPATCH p_scossl_rsa_signature_functions[] = {
     {OSSL_FUNC_SIGNATURE_NEWCTX, (void (*)(void))p_scossl_rsa_newctx},
     {OSSL_FUNC_SIGNATURE_DUPCTX, (void (*)(void))p_scossl_rsa_dupctx},
@@ -1410,74 +1168,6 @@ const OSSL_DISPATCH p_scossl_rsa_signature_functions[] = {
     {OSSL_FUNC_SIGNATURE_SET_CTX_MD_PARAMS, (void (*)(void))p_scossl_rsa_set_ctx_md_params},
     {OSSL_FUNC_SIGNATURE_SETTABLE_CTX_MD_PARAMS, (void (*)(void))p_scossl_rsa_settable_ctx_md_params},
     {0, NULL}};
-
-static const char **p_scossl_rsa_sigalg_query_key_types(void)
-{
-    static const char *keyTypes[] = { "RSA", NULL };
-    return keyTypes;
-}
-
-#define IMPLEMENT_SCOSSL_RSA_SIGALG(md, MD)                                                                         \
-    static SCOSSL_STATUS p_scossl_rsa_##md##_sign_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx,                            \
-                                                       _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,                        \
-                                                       _In_ const OSSL_PARAM params[])                              \
-    {                                                                                                               \
-        return p_scossl_rsa_sigalg_signverify_init(ctx, keyCtx, params, EVP_PKEY_OP_SIGN, MD);                      \
-    }                                                                                                               \
-                                                                                                                    \
-    static SCOSSL_STATUS p_scossl_rsa_##md##_verify_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx,                          \
-                                                         _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,                      \
-                                                         _In_ const OSSL_PARAM params[])                            \
-    {                                                                                                               \
-        return p_scossl_rsa_sigalg_signverify_init(ctx, keyCtx, params, EVP_PKEY_OP_VERIFY, MD);                    \
-    }                                                                                                               \
-                                                                                                                    \
-    static SCOSSL_STATUS p_scossl_rsa_##md##_sign_message_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx,                    \
-                                                               _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,                \
-                                                               _In_ const OSSL_PARAM params[])                      \
-    {                                                                                                               \
-        return p_scossl_rsa_sigalg_message_signverify_init(ctx, keyCtx, params, EVP_PKEY_OP_SIGNMSG, MD);           \
-    }                                                                                                               \
-                                                                                                                    \
-    static SCOSSL_STATUS p_scossl_rsa_##md##_verify_message_init(_Inout_ SCOSSL_RSA_SIGN_CTX *ctx,                  \
-                                                                 _In_ SCOSSL_PROV_RSA_KEY_CTX *keyCtx,              \
-                                                                 _In_ const OSSL_PARAM params[])                    \
-    {                                                                                                               \
-        return p_scossl_rsa_sigalg_message_signverify_init(ctx, keyCtx, params, EVP_PKEY_OP_VERIFYMSG, MD);         \
-    }                                                                                                               \
-                                                                                                                    \
-    const OSSL_DISPATCH p_scossl_rsa_##md##_signature_functions[] = {                                               \
-        {OSSL_FUNC_SIGNATURE_NEWCTX,               (void (*)(void))p_scossl_rsa_newctx},                            \
-        {OSSL_FUNC_SIGNATURE_DUPCTX,               (void (*)(void))p_scossl_rsa_dupctx},                            \
-        {OSSL_FUNC_SIGNATURE_FREECTX,              (void (*)(void))p_scossl_rsa_freectx},                           \
-        {OSSL_FUNC_SIGNATURE_SIGN_INIT,            (void (*)(void))p_scossl_rsa_##md##_sign_init},                  \
-        {OSSL_FUNC_SIGNATURE_SIGN,                 (void (*)(void))p_scossl_rsa_sign},                              \
-        {OSSL_FUNC_SIGNATURE_SIGN_MESSAGE_INIT,    (void (*)(void))p_scossl_rsa_##md##_sign_message_init},          \
-        {OSSL_FUNC_SIGNATURE_SIGN_MESSAGE_UPDATE,  (void (*)(void))p_scossl_rsa_sigalg_signverify_message_update},  \
-        {OSSL_FUNC_SIGNATURE_SIGN_MESSAGE_FINAL,   (void (*)(void))p_scossl_rsa_sigalg_sign_message_final},         \
-        {OSSL_FUNC_SIGNATURE_VERIFY_INIT,          (void (*)(void))p_scossl_rsa_##md##_verify_init},                \
-        {OSSL_FUNC_SIGNATURE_VERIFY,               (void (*)(void))p_scossl_rsa_verify},                            \
-        {OSSL_FUNC_SIGNATURE_VERIFY_MESSAGE_INIT,  (void (*)(void))p_scossl_rsa_##md##_verify_message_init},        \
-        {OSSL_FUNC_SIGNATURE_VERIFY_MESSAGE_UPDATE,(void (*)(void))p_scossl_rsa_sigalg_signverify_message_update},  \
-        {OSSL_FUNC_SIGNATURE_VERIFY_MESSAGE_FINAL, (void (*)(void))p_scossl_rsa_sigalg_verify_message_final},       \
-        {OSSL_FUNC_SIGNATURE_QUERY_KEY_TYPES,      (void (*)(void))p_scossl_rsa_sigalg_query_key_types},            \
-        {OSSL_FUNC_SIGNATURE_GET_CTX_PARAMS,       (void (*)(void))p_scossl_rsa_get_ctx_params},                    \
-        {OSSL_FUNC_SIGNATURE_GETTABLE_CTX_PARAMS,  (void (*)(void))p_scossl_rsa_gettable_ctx_params},               \
-        {OSSL_FUNC_SIGNATURE_SET_CTX_PARAMS,       (void (*)(void))p_scossl_rsa_sigalg_set_ctx_params},             \
-        {OSSL_FUNC_SIGNATURE_SETTABLE_CTX_PARAMS,  (void (*)(void))p_scossl_rsa_sigalg_settable_ctx_params},        \
-        {0, NULL}};
-
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha1,       OSSL_DIGEST_NAME_SHA1)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha224,     OSSL_DIGEST_NAME_SHA2_224)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha256,     OSSL_DIGEST_NAME_SHA2_256)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha384,     OSSL_DIGEST_NAME_SHA2_384)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha512,     OSSL_DIGEST_NAME_SHA2_512)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha512_224, OSSL_DIGEST_NAME_SHA2_512_224)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha512_256, OSSL_DIGEST_NAME_SHA2_512_256)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha3_224,   OSSL_DIGEST_NAME_SHA3_224)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha3_256,   OSSL_DIGEST_NAME_SHA3_256)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha3_384,   OSSL_DIGEST_NAME_SHA3_384)
-IMPLEMENT_SCOSSL_RSA_SIGALG(sha3_512,   OSSL_DIGEST_NAME_SHA3_512)
 
 #ifdef __cplusplus
 }
